requirements.txt
astroid==2.4.2
azure-functions==1.5.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pandas
pylint==2.6.0
pyodbc==4.0.30
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.2
wrapt==1.12.1
